Address PR review comments
- Remove Action from CommandType enum (it's a WPF adapter, not a command type)
- Change type discriminator from "type" to "\$type" on both LayoutElementDto and
  RawLayoutElementDto to avoid conflict with the behavioral Type property on
  CommandDefinitionDto and RawCommandDefinitionDto
- Move RawLayout into RawLayoutElementDto.cs, CompiledLayout into LayoutElementDto.cs,
  PreviewLayout into its own PreviewLayout.cs; delete Layouts.cs
